curriculum-modal moved
