--- src/app/curriculum-modal/curriculum-modal-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from "@angular/core";
-import { Routes, RouterModule } from "@angular/router";
-
-import { CurriculumModalPage } from "./curriculum-modal.page";
-
-const routes: Routes = [
-    {
-        path: "",
-        component: CurriculumModalPage,
-    },
-];
-
-@NgModule({
-    imports: [RouterModule.forChild(routes)],
-    exports: [RouterModule],
-})
-export class CurriculumModalPageRoutingModule {}
