+++ src/app/logging-form/curriculum-modal/curriculum-modal.module.ts
@@ -4,21 +4,12 @@ import { FormsModule } from "@angular/forms";
 
 import { IonicModule } from "@ionic/angular";
 
-import { CurriculumModalPageRoutingModule } from "./curriculum-modal-routing.module";
-
 import { CurriculumModalPage } from "./curriculum-modal.page";
-import { ComponentsModule } from "../_components/components.module";
+import { ComponentsModule } from "../../_components/components.module";
 import { TranslateModule } from "@ngx-translate/core";
 
 @NgModule({
-    imports: [
-        CommonModule,
-        FormsModule,
-        IonicModule,
-        ComponentsModule,
-        CurriculumModalPageRoutingModule,
-        TranslateModule,
-    ],
+    imports: [CommonModule, FormsModule, IonicModule, ComponentsModule, TranslateModule],
     declarations: [CurriculumModalPage],
 })
 export class CurriculumModalPageModule {}
