+++ src/app/logging-form/curriculum-modal/curriculum-modal.page.scss
